Moved settings to ts file to avoid experimental import warnings, clean up imports, removed dead code.
