+++ samples/upjf/tsconfig.json
@@ -2,7 +2,6 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "outDir": "js",
-        //"rootDir": "."
     },
     "include": [
         "./src/*.ts"
